Add files via upload
This project presents the development of a modern and user-friendly frontend interface for the website of Olympus Hotel, a fictional hospitality establishment. The aim of this project is to enhance the online presence of Olympus Hotel by providing a visually appealing and intuitive interface that caters to the needs of potential guests. Leveraging contemporary web development technologies such as HTML5, CSS3, and JavaScript, the frontend design focuses on seamless navigation, engaging visuals, and responsive layout to ensure compatibility across various devices and screen sizes. The project incorporates features such as interactive booking forms, photo galleries showcasing the hotel amenities, and informative sections highlighting the hotel's services and attractions.